# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GiocatoreRuolo = typeof GiocatoreRuolo[keyof typeof GiocatoreRuolo];


export const GiocatoreRuolo = {
  Attaccante: 'Attaccante',
  Difensore: 'Difensore',
  Centrocampista: 'Centrocampista',
  Portiere: 'Portiere',
} as const;

export interface Giocatore {
  id: number;
  nome: string;
  ruolo: GiocatoreRuolo;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minimum 0 */
  presenze: number;
  /** @minimum 0 */
  vittorie: number;
}

export type GiocatoreInputRuolo = typeof GiocatoreInputRuolo[keyof typeof GiocatoreInputRuolo];


export const GiocatoreInputRuolo = {
  Attaccante: 'Attaccante',
  Difensore: 'Difensore',
  Centrocampista: 'Centrocampista',
  Portiere: 'Portiere',
} as const;

export interface GiocatoreInput {
  /** @minLength 1 */
  nome: string;
  ruolo: GiocatoreInputRuolo;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}

export type GiocatoreUpdateRuolo = typeof GiocatoreUpdateRuolo[keyof typeof GiocatoreUpdateRuolo];


export const GiocatoreUpdateRuolo = {
  Attaccante: 'Attaccante',
  Difensore: 'Difensore',
  Centrocampista: 'Centrocampista',
  Portiere: 'Portiere',
} as const;

export interface GiocatoreUpdate {
  /** @minLength 1 */
  nome?: string;
  ruolo?: GiocatoreUpdateRuolo;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
}

export interface BulkDeleteInput {
  /** @minItems 1 */
  ids: number[];
}

export interface BulkDeleteResult {
  deleted: number;
}

export interface Stats {
  totaleGiocatori: number;
  totalePresenze: number;
  totaleVittorie: number;
  /** @nullable */
  miglioreAttacco: string | null;
  /** @nullable */
  migliorePortiere: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface GeneraSquadreInput {
  /** @minItems 3 */
  giocatoriIds: number[];
  secondaPartitaIds?: number[];
}

export type GiocatoreConLivelloRuolo = typeof GiocatoreConLivelloRuolo[keyof typeof GiocatoreConLivelloRuolo];


export const GiocatoreConLivelloRuolo = {
  Attaccante: 'Attaccante',
  Difensore: 'Difensore',
  Centrocampista: 'Centrocampista',
  Portiere: 'Portiere',
  Virtuale: 'Virtuale',
} as const;

export interface GiocatoreConLivello {
  id: number;
  nome: string;
  ruolo: GiocatoreConLivelloRuolo;
  rating: number;
  presenze: number;
  vittorie: number;
  livello: number;
  virtuale?: boolean;
}

export interface Squadra {
  nome: string;
  giocatori: GiocatoreConLivello[];
  livelleTotale: number;
}

export interface GeneraSquadreResult {
  squadraA: Squadra;
  squadraB: Squadra;
  squadraC: Squadra;
}