# lib/api-zod/src/generated/types/bulkDeleteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */

export interface BulkDeleteInput {
  /** @minItems 1 */
  ids: number[];
}