# lib/api-zod/src/generated/types/generaSquadreResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
import type { Squadra } from './squadra';

export interface GeneraSquadreResult {
  squadraA: Squadra;
  squadraB: Squadra;
  squadraC: Squadra;
}