# lib/api-zod/src/generated/types/giocatore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
import type { GiocatoreRuolo } from './giocatoreRuolo';

export interface Giocatore {
  id: number;
  nome: string;
  ruolo: GiocatoreRuolo;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minimum 0 */
  presenze: number;
  /** @minimum 0 */
  vittorie: number;
}