# lib/api-zod/src/generated/types/giocatoreConLivello.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
import type { GiocatoreConLivelloRuolo } from './giocatoreConLivelloRuolo';

export interface GiocatoreConLivello {
  id: number;
  nome: string;
  ruolo: GiocatoreConLivelloRuolo;
  rating: number;
  presenze: number;
  vittorie: number;
  livello: number;
  virtuale?: boolean;
}