# lib/api-zod/src/generated/types/giocatoreInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
import type { GiocatoreInputRuolo } from './giocatoreInputRuolo';

export interface GiocatoreInput {
  /** @minLength 1 */
  nome: string;
  ruolo: GiocatoreInputRuolo;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}