# lib/api-zod/src/generated/types/giocatoreUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
import type { GiocatoreUpdateRuolo } from './giocatoreUpdateRuolo';

export interface GiocatoreUpdate {
  /** @minLength 1 */
  nome?: string;
  ruolo?: GiocatoreUpdateRuolo;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
}