# lib/api-zod/src/generated/types/squadra.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Torneo Calcetto
 * OpenAPI spec version: 0.1.0
 */
import type { GiocatoreConLivello } from './giocatoreConLivello';

export interface Squadra {
  nome: string;
  giocatori: GiocatoreConLivello[];
  livelleTotale: number;
}